+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rules Component.
 - Logo Component.
 - Intro Component.
+- Authors Component. 
 - Background styles.
 - Background dice component.
 - Style reset.